requirements.txt
certifi==2019.11.28
Django==2.2.5
djangorestframework==3.10.3
pytz==2019.3
sqlparse==0.3.0
wincertstore==0.2
